refactor(404-page): move "404-page" folder to /shared
Refactor in folder structure to follow the guide

Closes #8
