+++ src/app/app.module.ts
@@ -14,7 +14,7 @@ import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
 import { LoginComponent } from './login/login.component';
 import { HeaderComponent } from './header/header.component';
-import { NothingFoundComponent } from './nothing-found/nothing-found.component';
+import { NothingFoundComponent } from './shared/nothing-found/nothing-found.component';
 import { RegisterComponent } from './shared/register/register.component';
 import { ExpenseListComponent } from './expenses/expense-list/expense-list.component';
 import { CategoryListComponent } from './categories/category-list/category-list.component';
